fixed issue with DI and pushed response body logic in middleware class

+++ Core/Message/Response/NegotiateResponse.cs
@@ -4,13 +4,13 @@ namespace Sahurjt.Signalr.Dashboard.Core.Message.Response
     internal class NegotiateResponse 
     {
         public string ConnectionId { get; set; }
-        public int ConnectionTimeout { get; set; }
+        public double ConnectionTimeout { get; set; }
         public string ConnectionToken { get; set; }
-        public int DisconnectTimeout { get; set; }
-        public int KeepAliveTimeout { get; set; }
-        public int LongPollDelay { get; set; }
+        public double DisconnectTimeout { get; set; }
+        public double KeepAliveTimeout { get; set; }
+        public double LongPollDelay { get; set; }
         public string ProtocolVersion { get; set; }
-        public int TransportConnectTime { get; set; }
+        public double TransportConnectTime { get; set; }
         public bool TryWebSockets { get; set; }
         public string Url { get; set; }
     }
